fixed dtype error for 'num-anchors-matched'; misc bugs

+++ retinanet/trainer/trainer.py
@@ -263,7 +263,7 @@ class Trainer:
         self.optimizer.apply_gradients(
             zip(gradients, self._model.trainable_variables))
 
-        loss['num-anchors-matched'] /= tf.shape(images)[0]
+        loss['num-anchors-matched'] /= tf.cast(tf.shape(images)[0], dtype=tf.float32)
         loss['gradient-norm'] = tf.linalg.global_norm(gradients) * self.num_replicas
         return loss
